Plugin Loading and App Structure
This commit starts on the plugin loader which will be
responsible for loading the plugins as defined in the
app config file. Additional work and testing is still
needed in this area.

Additionally, the main app has been updated with a
basic structure which will define how it runs. Again,
this will need to be completed and will likely need
some tweaking (eg. for "interrupts"), but it gives
some direction.